Endpoint findAll and save

+++ dev/src/main/resources/db/migration/V1__create_table_problema_saude.sql
@@ -1,5 +1,5 @@
 CREATE TABLE problema_saude (
-    id_problema BIGINT PRIMARY KEY,
+    id_problema SERIAL PRIMARY KEY,
     nome VARCHAR(50) NOT NULL,
     grau_do_problema BIGINT NOT NULL
 );
